Ported test project's code for proof UI
This is probably the biggest single commit so far. It incorporates about 2 weeks of work on standalone test projects. I wanted a clean sandbox to test all these UI features and didn't want to mess up any of EGG's design so far. Now that I've mostly satisfied with it, I've brought it in. Nothing is hooked up yet, but the bare minimum API to connect everything is in place.