Very first modification command
Doesn't actually display to UI yet, but entering modification mode with / (slash key) and hitting 'w' will attempt to surround the selection/highlight with a double cut. This produces a new StepItem that can be clicked. Clicking it shows that the tree state is indeed correct as it prints to the console.